+++ server/server_test.go
@@ -163,8 +163,6 @@ func (suite *ServerTestSuite) TestIssueRedeemV2() {
 	_, err = suite.srv.db.Query(`UPDATE v3_issuers SET expires_at=$1 WHERE issuer_id=$2`, time.Now().AddDate(0, 0, -1), issuer.ID)
 	suite.Require().NoError(err, "failed to expire issuer")
 	// keys are what rotate now, not the issuer itself
-	// issuers, _ := suite.srv.fetchIssuers(issuerType)
-	// suite.Assert().Equal(len(*issuers), 2, "There should be two issuers of same type")
 	issuer, _ = suite.srv.GetLatestIssuer(issuerType, issuerCohort)
 
 	resp, err = suite.attemptRedeem(server.URL, preimageText2, sigText2, issuerType, msg)
@@ -234,10 +232,6 @@ func (suite *ServerTestSuite) TestNewIssueRedeemV2() {
 	_, err = suite.srv.db.Query(`UPDATE v3_issuers SET expires_at=$1 WHERE issuer_id=$2`, time.Now().AddDate(0, 0, -1), issuer.ID)
 	suite.Require().NoError(err, "failed to expire issuer")
 
-	// no more duplicate issuers for rotations, the keys will be rotated
-	//issuers, _ := suite.srv.fetchIssuers(issuerType)
-	//suite.Assert().Equal(len(*issuers), 2, "There should be two issuers of same type")
-
 	resp, err = suite.attemptRedeem(server.URL, preimageText2, sigText2, issuerType, msg)
 	suite.Assert().NoError(err, "HTTP Request should complete")
 	suite.Assert().Equal(http.StatusBadRequest, resp.StatusCode, "Expired Issuers should fail")
@@ -261,7 +255,7 @@ func (suite *ServerTestSuite) request(method string, URL string, payload io.Read
 	return http.DefaultClient.Do(req)
 }
 
-func (suite *ServerTestSuite) createIssuer(serverURL string, issuerType string, issuerCohort int) *crypto.PublicKey {
+func (suite *ServerTestSuite) createIssuer(serverURL string, issuerType string, issuerCohort int16) *crypto.PublicKey {
 	payload := fmt.Sprintf(`{"name":"%s", "cohort": %d, "max_tokens":100}`, issuerType, issuerCohort)
 	createIssuerURL := fmt.Sprintf("%s/v1/issuer/", serverURL)
 	resp, err := suite.request("POST", createIssuerURL, bytes.NewBuffer([]byte(payload)))
@@ -309,7 +303,7 @@ func (suite *ServerTestSuite) getAllIssuers(serverURL string) []issuerResponse {
 	return issuerResp
 }
 
-func (suite *ServerTestSuite) createIssuerWithExpiration(serverURL string, issuerType string, issuerCohort int, expiresAt time.Time) *crypto.PublicKey {
+func (suite *ServerTestSuite) createIssuerWithExpiration(serverURL string, issuerType string, issuerCohort int16, expiresAt time.Time) *crypto.PublicKey {
 	payload := fmt.Sprintf(`{"name":"%s", "cohort": %d, "max_tokens":100, "expires_at":"%s"}`, issuerType, issuerCohort, expiresAt.Format("2006-01-02T15:04:05Z07:00"))
 	// v2+ has expirations
 	createIssuerURL := fmt.Sprintf("%s/v2/issuer/", serverURL)
@@ -403,11 +397,11 @@ func (suite *ServerTestSuite) attemptRedeem(serverURL string, preimageText []byt
 	return suite.request("POST", redeemURL, bytes.NewBuffer([]byte(payload)))
 }
 
-func (suite *ServerTestSuite) createCohortToken(serverURL string, issuerType string, issuerCohort int, publicKey *crypto.PublicKey) *crypto.UnblindedToken {
+func (suite *ServerTestSuite) createCohortToken(serverURL string, issuerType string, issuerCohort int16, publicKey *crypto.PublicKey) *crypto.UnblindedToken {
 	return suite.createCohortTokens(serverURL, issuerType, issuerCohort, publicKey, 1)[0]
 }
 
-func (suite *ServerTestSuite) createCohortTokens(serverURL string, issuerType string, issuerCohort int, publicKey *crypto.PublicKey, numTokens int) []*crypto.UnblindedToken {
+func (suite *ServerTestSuite) createCohortTokens(serverURL string, issuerType string, issuerCohort int16, publicKey *crypto.PublicKey, numTokens int) []*crypto.UnblindedToken {
 	tokens := make([]*crypto.Token, numTokens)
 	blindedTokens := make([]*crypto.BlindedToken, numTokens)
